[P073] Restructure code per display in separate files
